Refactor productivity screen to decouple view from logic
- Extract `useProductivityViewModel` custom hook to manage UI state, filtering, and "Today plan" business logic.
- Replace inline style tokens with dynamic `useTheme` mappings.
- Clean up the UI view, reducing lines of code dramatically.
- Ensure correct TS types and paths are used.
