+++ frontend/app/(main)/productivity.tsx
@@ -1,6 +1,5 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useMemo } from 'react';
 import {
-  Alert,
   FlatList,
   Platform,
   Pressable,
@@ -11,8 +10,6 @@ import {
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { Ionicons } from '@expo/vector-icons';
-import { useLocalSearchParams, usePathname, useRouter } from 'expo-router';
-import { useTranslation } from 'react-i18next';
 import { AppText } from '../../src/components/AppText';
 import { CreateNoteModal } from '../../src/components/productivity/CreateNoteModal';
 import { CreateTaskModal } from '../../src/components/productivity/CreateTaskModal';
@@ -20,323 +17,244 @@ import { NoteCard } from '../../src/components/productivity/NoteCard';
 import { TaskCard } from '../../src/components/productivity/TaskCard';
 import { theme } from '../../src/core/theme';
 import { CalendarEvent, Note, Task } from '../../src/core/models';
-import { useProductivityStore } from '../../src/store/useProductivityStore';
-import { DESIGN_TOKENS } from '@/core/design/tokens';
+import { useTheme } from '../../src/hooks/useTheme';
+import { RenderItemData, useProductivityViewModel } from '../../src/hooks/viewmodels/useProductivityViewModel';
 
-const T = {
-  background: { light: DESIGN_TOKENS.colors.pageBg },
-  surface: { light: DESIGN_TOKENS.colors.surface, highLight: DESIGN_TOKENS.colors.surfaceSoft },
-  border: { light: DESIGN_TOKENS.colors.border },
-  onSurface: {
-    light: DESIGN_TOKENS.colors.text,
-    mutedLight: DESIGN_TOKENS.colors.muted,
-    disabledLight: DESIGN_TOKENS.colors.faint,
-  },
-  primary: {
-    DEFAULT: DESIGN_TOKENS.colors.primary,
-    surfaceLight: DESIGN_TOKENS.colors.primarySoft,
-  },
-  white: '#FFFFFF',
-  transparent: 'transparent',
-};
 const S = theme.spacing;
 const R = theme.borderRadius;
 
-type Tab = 'today' | 'all' | 'notes' | 'tasks';
-type TaskPriority = 'all' | 'overdue' | 'today' | 'upcoming' | 'no_due';
-
-type RenderItemData = {
-  date?: number;
-  type: 'note' | 'task' | 'event';
-  item: Note | Task | CalendarEvent;
-  id: string;
-};
-
 export default function ProductivityScreen() {
-  const { t, i18n } = useTranslation();
-  const router = useRouter();
-  const pathname = usePathname();
-  const params = useLocalSearchParams() as Record<string, string | string[] | undefined>;
-  const openCreateTask = params.openCreateTask;
-  const openCreateNote = params.openCreateNote;
-  const [activeTab, setActiveTab] = useState<Tab>('today');
-  const [taskPriority, setTaskPriority] = useState<TaskPriority>('all');
-  const [searchQuery, setSearchQuery] = useState('');
-  const [showCreateNote, setShowCreateNote] = useState(false);
-  const [showCreateTask, setShowCreateTask] = useState(false);
-  const [todayPlan, setTodayPlan] = useState<string | null>(null);
-
   const {
-    notes,
-    tasks,
-    events,
+    t,
+    i18n,
+    activeTab,
+    setActiveTab,
+    taskPriority,
+    setTaskPriority,
+    searchQuery,
+    setSearchQuery,
+    showCreateNote,
+    setShowCreateNote,
+    showCreateTask,
+    setShowCreateTask,
+    todayPlan,
+    setTodayPlan,
+    isLoading,
+    handleRefresh,
+    confirmDelete,
+    pendingTasksCount,
+    taskPriorityCounts,
+    dataToRender,
+    generateTodayPlan,
     fetchNotes,
     fetchTasks,
-    fetchEvents,
-    isLoading,
     deleteNote,
     deleteTask,
     toggleTask,
-  } = useProductivityStore();
+  } = useProductivityViewModel();
 
-  useEffect(() => {
-    const controller = new AbortController();
-    fetchNotes(controller.signal);
-    fetchTasks(controller.signal);
-    fetchEvents(controller.signal);
-    return () => {
-      controller.abort();
-    };
-  }, [fetchEvents, fetchNotes, fetchTasks]);
+  const { C } = useTheme();
 
-  useEffect(() => {
-    if (openCreateTask === '1' || openCreateTask === 'true') {
-      setShowCreateTask(true);
-      const nextParams = Object.fromEntries(
-        Object.entries(params).filter(
-          ([key, value]) => key !== 'openCreateTask' && typeof value === 'string'
-        )
-      );
-      router.replace({ pathname, params: nextParams });
-    }
-  }, [openCreateTask, params, pathname, router]);
-
-  useEffect(() => {
-    if (openCreateNote === '1' || openCreateNote === 'true') {
-      setShowCreateNote(true);
-      const nextParams = Object.fromEntries(
-        Object.entries(params).filter(
-          ([key, value]) => key !== 'openCreateNote' && typeof value === 'string'
-        )
-      );
-      router.replace({ pathname, params: nextParams });
-    }
-  }, [openCreateNote, params, pathname, router]);
-
-  const handleRefresh = useCallback(() => {
-    fetchNotes();
-    fetchTasks();
-    fetchEvents();
-  }, [fetchEvents, fetchNotes, fetchTasks]);
-
-  const confirmDelete = useCallback(
-    (action: () => Promise<void>) =>
-      Alert.alert(
-        t('productivity.delete') || 'Delete',
-        t('productivity.are_you_sure') || 'Are you sure?',
-        [
-          { text: t('settings.actions.cancel') || 'Cancel', style: 'cancel' },
-          {
-            text: t('settings.actions.delete') || 'Delete',
-            style: 'destructive',
-            onPress: () => action(),
-          },
-        ]
-      ),
-    [t]
+  const styles = useMemo(
+    () =>
+      StyleSheet.create({
+        container: {
+          flex: 1,
+          backgroundColor: C.bg,
+        },
+        headerContainer: {
+          paddingHorizontal: S.xl,
+          paddingTop: S.md,
+          paddingBottom: S.lg,
+          backgroundColor: C.surface,
+          ...theme.elevation.sm,
+          zIndex: 10,
+        },
+        headerRow: {
+          flexDirection: 'row',
+          justifyContent: 'space-between',
+          alignItems: 'center',
+          marginBottom: S.lg,
+        },
+        headerTitle: {
+          color: C.text,
+          fontSize: 28,
+          fontWeight: '800',
+          letterSpacing: -0.5,
+        },
+        headerSubtitle: {
+          color: C.muted,
+          fontSize: 14,
+          marginTop: S.xs,
+        },
+        headerActions: {
+          flexDirection: 'row',
+          gap: S.sm,
+        },
+        iconButton: {
+          width: 40,
+          height: 40,
+          borderRadius: R.full,
+          backgroundColor: C.primarySurface,
+          alignItems: 'center',
+          justifyContent: 'center',
+        },
+        searchContainer: {
+          flexDirection: 'row',
+          alignItems: 'center',
+          backgroundColor: C.bg,
+          borderRadius: R.lg,
+          paddingHorizontal: S.md,
+          height: 44,
+          borderWidth: 1,
+          borderColor: C.border,
+        },
+        searchIcon: {
+          marginRight: S.sm,
+        },
+        searchInput: {
+          flex: 1,
+          color: C.text,
+          fontSize: 16,
+          height: '100%',
+        },
+        tabsContainer: {
+          flexDirection: 'row',
+          backgroundColor: C.surfaceHigh,
+          borderRadius: R.xl,
+          padding: S.xs,
+          marginHorizontal: S.xl,
+          marginTop: S.lg,
+          marginBottom: S.md,
+          borderWidth: 1,
+          borderColor: C.border,
+        },
+        tabButton: {
+          flex: 1,
+          paddingVertical: S.sm,
+          alignItems: 'center',
+          borderRadius: R.lg,
+          backgroundColor: 'transparent',
+        },
+        tabButtonActive: {
+          backgroundColor: C.surface,
+          ...theme.elevation.sm,
+        },
+        tabText: {
+          fontSize: 14,
+          fontWeight: '500',
+          color: C.muted,
+        },
+        tabTextActive: {
+          fontWeight: '700',
+          color: C.text,
+        },
+        listContent: {
+          paddingHorizontal: S.xl,
+          paddingBottom: 100,
+          paddingTop: S.sm,
+        },
+        eventCard: {
+          flexDirection: 'row',
+          alignItems: 'center',
+          backgroundColor: C.surface,
+          borderWidth: 1,
+          borderColor: C.border,
+          borderRadius: R.xl,
+          padding: S.lg,
+          marginBottom: S.md,
+          ...theme.elevation.sm,
+        },
+        eventIcon: {
+          width: 32,
+          height: 32,
+          borderRadius: R.lg,
+          backgroundColor: C.primarySurface,
+          alignItems: 'center',
+          justifyContent: 'center',
+          marginRight: S.md,
+        },
+        eventBody: {
+          flex: 1,
+        },
+        eventTitle: {
+          color: C.text,
+          fontWeight: '700',
+          fontSize: 15,
+        },
+        eventMeta: {
+          color: C.muted,
+          marginTop: 2,
+          fontSize: 13,
+        },
+        emptyContainer: {
+          alignItems: 'center',
+          paddingVertical: S.massive,
+        },
+        emptyIconCircle: {
+          width: 80,
+          height: 80,
+          borderRadius: 40,
+          backgroundColor: C.primarySurface,
+          alignItems: 'center',
+          justifyContent: 'center',
+          marginBottom: S.lg,
+        },
+        emptyTitle: {
+          marginBottom: S.sm,
+          textAlign: 'center',
+          color: C.text,
+        },
+        emptySubtitle: {
+          textAlign: 'center',
+          marginBottom: S.xl,
+          color: C.muted,
+          paddingHorizontal: S.xxxl,
+          lineHeight: 22,
+        },
+        emptyActions: {
+          flexDirection: 'row',
+          gap: S.md,
+        },
+        emptyButton: {
+          flexDirection: 'row',
+          alignItems: 'center',
+          backgroundColor: C.primary,
+          borderRadius: R.xl,
+          paddingVertical: S.md,
+          paddingHorizontal: S.xl,
+          ...theme.elevation.md,
+        },
+        emptyButtonSecondary: {
+          backgroundColor: C.primarySurface,
+          ...Platform.select({
+            ios: {
+              shadowOpacity: 0,
+              elevation: 0,
+            },
+            android: {
+              elevation: 0,
+            },
+            default: {
+              elevation: 0,
+            },
+          }),
+        },
+        emptyButtonText: {
+          color: '#FFFFFF',
+          fontWeight: '700',
+          fontSize: 15,
+        },
+        emptyButtonTextSecondary: {
+          color: C.primary,
+          fontWeight: '700',
+          fontSize: 15,
+        },
+      }),
+    [C]
   );
 
-  const filteredNotes = useMemo(() => {
-    if (!searchQuery) return notes;
-    const lowerQ = searchQuery.toLowerCase();
-    return notes.filter(
-      (n) => n.title.toLowerCase().includes(lowerQ) || n.content.toLowerCase().includes(lowerQ)
-    );
-  }, [notes, searchQuery]);
-
-  const filteredTasks = useMemo(() => {
-    if (!searchQuery) return tasks;
-    const lowerQ = searchQuery.toLowerCase();
-    return tasks.filter(
-      (task) =>
-        task.title.toLowerCase().includes(lowerQ) ||
-        (task.description?.toLowerCase().includes(lowerQ) ?? false)
-    );
-  }, [searchQuery, tasks]);
-
-  const filteredEvents = useMemo(() => {
-    if (!searchQuery) return events;
-    const lowerQ = searchQuery.toLowerCase();
-    return events.filter(
-      (event) =>
-        event.title.toLowerCase().includes(lowerQ) ||
-        (event.description?.toLowerCase().includes(lowerQ) ?? false) ||
-        (event.location?.toLowerCase().includes(lowerQ) ?? false)
-    );
-  }, [events, searchQuery]);
-
-  const pendingTasksCount = useMemo(
-    () => filteredTasks.filter((task) => !task.completed).length,
-    [filteredTasks]
-  );
-
-  const getTaskPriority = useCallback((task: Task): Exclude<TaskPriority, 'all'> => {
-    if (!task.due_date) return 'no_due';
-    const now = new Date();
-    const due = new Date(task.due_date);
-    if (isNaN(due.getTime())) return 'no_due';
-    const dayStart = new Date(now);
-    dayStart.setHours(0, 0, 0, 0);
-    const tomorrow = new Date(dayStart);
-    tomorrow.setDate(tomorrow.getDate() + 1);
-    if (due < dayStart) return 'overdue';
-    if (due >= dayStart && due < tomorrow) return 'today';
-    return 'upcoming';
-  }, []);
-
-  const taskPriorityCounts = useMemo(() => {
-    const counts: Record<TaskPriority, number> = {
-      all: 0,
-      overdue: 0,
-      today: 0,
-      upcoming: 0,
-      no_due: 0,
-    };
-    filteredTasks
-      .filter((task) => !task.completed)
-      .forEach((task) => {
-        counts.all += 1;
-        counts[getTaskPriority(task)] += 1;
-      });
-    return counts;
-  }, [filteredTasks, getTaskPriority]);
-
-  const prioritizedTasks = useMemo(() => {
-    const tasksToRank = filteredTasks.filter((task) => !task.completed);
-    const pool =
-      taskPriority === 'all'
-        ? tasksToRank
-        : tasksToRank.filter((task) => getTaskPriority(task) === taskPriority);
-    return [...pool].sort((a, b) => {
-      const aDue = a.due_date ? new Date(a.due_date).getTime() : Number.MAX_SAFE_INTEGER;
-      const bDue = b.due_date ? new Date(b.due_date).getTime() : Number.MAX_SAFE_INTEGER;
-      return aDue - bDue;
-    });
-  }, [filteredTasks, getTaskPriority, taskPriority]);
-
-  const mixedData = useMemo(() => {
-    const data: RenderItemData[] = [];
-    filteredNotes.forEach((note) => {
-      data.push({
-        type: 'note',
-        item: note,
-        id: `note-${note.id}`,
-        date: new Date(note.created_at).getTime(),
-      });
-    });
-    filteredTasks.forEach((task) => {
-      data.push({
-        type: 'task',
-        item: task,
-        id: `task-${task.id}`,
-        date: new Date(task.created_at).getTime(),
-      });
-    });
-    return data.sort((a, b) => (b.date || 0) - (a.date || 0));
-  }, [filteredNotes, filteredTasks]);
-
-  const todayData = useMemo(() => {
-    const now = new Date();
-    const start = new Date(now);
-    start.setHours(0, 0, 0, 0);
-    const end = new Date(start);
-    end.setDate(end.getDate() + 1);
-    const weekEnd = new Date(start);
-    weekEnd.setDate(weekEnd.getDate() + 7);
-
-    const overdueTasks = filteredTasks.filter((task) => {
-      if (task.completed || !task.due_date) return false;
-      return new Date(task.due_date) < start;
-    });
-
-    const dueTodayTasks = filteredTasks.filter((task) => {
-      if (task.completed || !task.due_date) return false;
-      const dueDate = new Date(task.due_date);
-      return dueDate >= start && dueDate < end;
-    });
-
-    const upcomingEvents = filteredEvents.filter((event) => {
-      const startTime = new Date(event.start_time);
-      return startTime >= start && startTime < weekEnd;
-    });
-
-    const items: RenderItemData[] = [
-      ...overdueTasks.map((task) => ({
-        type: 'task' as const,
-        item: task,
-        id: `overdue-${task.id}`,
-        date: task.due_date ? new Date(task.due_date).getTime() : undefined,
-      })),
-      ...dueTodayTasks.map((task) => ({
-        type: 'task' as const,
-        item: task,
-        id: `today-${task.id}`,
-        date: task.due_date ? new Date(task.due_date).getTime() : undefined,
-      })),
-      ...upcomingEvents.map((event) => ({
-        type: 'event' as const,
-        item: event,
-        id: `event-${event.id}`,
-        date: new Date(event.start_time).getTime(),
-      })),
-    ];
-
-    return items.sort((a, b) => (a.date || 0) - (b.date || 0));
-  }, [filteredEvents, filteredTasks]);
-
-  const dataToRender: RenderItemData[] =
-    activeTab === 'today'
-      ? todayData.filter((entry) => {
-          if (entry.type !== 'task') return true;
-          if (taskPriority === 'all') return true;
-          return getTaskPriority(entry.item as Task) === taskPriority;
-        })
-      : activeTab === 'all'
-        ? mixedData
-        : activeTab === 'notes'
-          ? filteredNotes.map((note) => ({ type: 'note' as const, item: note, id: note.id }))
-          : prioritizedTasks.map((task) => ({ type: 'task' as const, item: task, id: task.id }));
-
-  const generateTodayPlan = useCallback(() => {
-    const now = new Date();
-    const nextTasks = prioritizedTasks.slice(0, 3);
-    const nextEvents = [...filteredEvents]
-      .filter((event) => new Date(event.end_time) >= now)
-      .sort((a, b) => new Date(a.start_time).getTime() - new Date(b.start_time).getTime())
-      .slice(0, 3);
-
-    const lines: string[] = [];
-    if (taskPriorityCounts.overdue > 0) {
-      lines.push(`1) Recover overdue: start with ${taskPriorityCounts.overdue} overdue task(s).`);
-    } else {
-      lines.push('1) No overdue tasks: start with highest-impact open work.');
-    }
-
-    if (nextTasks.length > 0) {
-      lines.push(`2) Focus block: ${nextTasks.map((task) => task.title).join(', ')}.`);
-    } else {
-      lines.push('2) Focus block: no pending tasks, use this for planning or review.');
-    }
-
-    if (nextEvents.length > 0) {
-      const eventLine = nextEvents
-        .map((event) =>
-          new Intl.DateTimeFormat(i18n.language, {
-            hour: '2-digit',
-            minute: '2-digit',
-          }).format(new Date(event.start_time))
-        )
-        .join(', ');
-      lines.push(`3) Calendar checkpoints at ${eventLine}.`);
-    } else {
-      lines.push('3) Calendar is light: reserve time for deep work and wrap-up.');
-    }
-
-    setTodayPlan(lines.join('\n'));
-    setActiveTab('today');
-  }, [filteredEvents, i18n.language, prioritizedTasks, taskPriorityCounts.overdue]);
-
-  const renderItem = useCallback(
+  const renderItem = React.useCallback(
     ({ item }: { item: RenderItemData }) => {
       if (item.type === 'note') {
         const note = item.item as Note;
@@ -347,7 +265,7 @@ export default function ProductivityScreen() {
         return (
           <View style={styles.eventCard}>
             <View style={styles.eventIcon}>
-              <Ionicons name="calendar-outline" size={16} color={T.primary.DEFAULT} />
+              <Ionicons name="calendar-outline" size={16} color={C.primary} />
             </View>
             <View style={styles.eventBody}>
               <AppText style={styles.eventTitle}>{event.title}</AppText>
@@ -374,7 +292,7 @@ export default function ProductivityScreen() {
         />
       );
     },
-    [confirmDelete, deleteNote, deleteTask, i18n.language, toggleTask]
+    [confirmDelete, deleteNote, deleteTask, i18n.language, toggleTask, C.primary, styles]
   );
 
   return (
@@ -398,35 +316,30 @@ export default function ProductivityScreen() {
               onPress={generateTodayPlan}
               style={({ pressed }) => [styles.iconButton, pressed && { opacity: 0.7 }]}
             >
-              <Ionicons name="sparkles" size={20} color={T.primary.DEFAULT} />
+              <Ionicons name="sparkles" size={20} color={C.primary} />
             </Pressable>
             <Pressable
               onPress={() => setShowCreateNote(true)}
               style={({ pressed }) => [styles.iconButton, pressed && { opacity: 0.7 }]}
             >
-              <Ionicons name="document-text" size={20} color={T.primary.DEFAULT} />
+              <Ionicons name="document-text" size={20} color={C.primary} />
             </Pressable>
             <Pressable
               onPress={() => setShowCreateTask(true)}
               style={({ pressed }) => [styles.iconButton, pressed && { opacity: 0.7 }]}
             >
-              <Ionicons name="checkbox" size={20} color={T.primary.DEFAULT} />
+              <Ionicons name="checkbox" size={20} color={C.primary} />
             </Pressable>
           </View>
         </View>
 
         <View style={styles.searchContainer}>
-          <Ionicons
-            name="search"
-            size={18}
-            color={T.onSurface.mutedLight}
-            style={styles.searchIcon}
-          />
+          <Ionicons name="search" size={18} color={C.muted} style={styles.searchIcon} />
           <TextInput
             value={searchQuery}
             onChangeText={setSearchQuery}
             placeholder={t('productivity.search') || 'Search notes & tasks...'}
-            placeholderTextColor={T.onSurface.disabledLight}
+            placeholderTextColor={C.faint}
             style={styles.searchInput}
             clearButtonMode="while-editing"
           />
@@ -468,7 +381,10 @@ export default function ProductivityScreen() {
         >
           {(
             [
-              { key: 'all', label: t('productivity.priority.all', { count: taskPriorityCounts.all }) },
+              {
+                key: 'all',
+                label: t('productivity.priority.all', { count: taskPriorityCounts.all }),
+              },
               {
                 key: 'overdue',
                 label: t('productivity.priority.overdue', { count: taskPriorityCounts.overdue }),
@@ -494,9 +410,8 @@ export default function ProductivityScreen() {
                 {
                   borderRadius: 12,
                   borderWidth: 1,
-                  borderColor: taskPriority === option.key ? T.primary.DEFAULT : T.border.light,
-                  backgroundColor:
-                    taskPriority === option.key ? T.primary.surfaceLight : T.surface.light,
+                  borderColor: taskPriority === option.key ? C.primary : C.border,
+                  backgroundColor: taskPriority === option.key ? C.primarySurface : C.surface,
                   paddingHorizontal: 10,
                   paddingVertical: 6,
                   marginRight: 8,
@@ -508,7 +423,7 @@ export default function ProductivityScreen() {
               <AppText
                 variant="caption"
                 style={{
-                  color: taskPriority === option.key ? T.primary.DEFAULT : T.onSurface.mutedLight,
+                  color: taskPriority === option.key ? C.primary : C.muted,
                   fontWeight: '700',
                 }}
               >
@@ -528,7 +443,7 @@ export default function ProductivityScreen() {
           <RefreshControl
             refreshing={isLoading && dataToRender.length > 0}
             onRefresh={handleRefresh}
-            tintColor={T.primary.DEFAULT}
+            tintColor={C.primary}
           />
         }
         renderItem={renderItem}
@@ -537,9 +452,9 @@ export default function ProductivityScreen() {
             <View
               style={{
                 borderRadius: R.xl,
-                backgroundColor: T.primary.surfaceLight,
+                backgroundColor: C.primarySurface,
                 borderWidth: 1,
-                borderColor: T.border.light,
+                borderColor: C.border,
                 padding: S.lg,
                 marginBottom: S.md,
               }}
@@ -551,14 +466,14 @@ export default function ProductivityScreen() {
                   alignItems: 'center',
                 }}
               >
-                <AppText style={{ color: T.onSurface.light, fontWeight: '700', fontSize: 15 }}>
+                <AppText style={{ color: C.text, fontWeight: '700', fontSize: 15 }}>
                   Today plan
                 </AppText>
                 <Pressable onPress={() => setTodayPlan(null)}>
-                  <Ionicons name="close" size={16} color={T.onSurface.mutedLight} />
+                  <Ionicons name="close" size={16} color={C.muted} />
                 </Pressable>
               </View>
-              <AppText style={{ color: T.onSurface.mutedLight, marginTop: 8, lineHeight: 20 }}>
+              <AppText style={{ color: C.muted, marginTop: 8, lineHeight: 20 }}>
                 {todayPlan}
               </AppText>
             </View>
@@ -578,7 +493,7 @@ export default function ProductivityScreen() {
                         : 'planet'
                 }
                 size={42}
-                color={T.primary.DEFAULT}
+                color={C.primary}
               />
             </View>
             <AppText variant="h3" style={styles.emptyTitle}>
@@ -608,7 +523,7 @@ export default function ProductivityScreen() {
                     onPress={() => setShowCreateNote(true)}
                     style={({ pressed }) => [styles.emptyButton, pressed && { opacity: 0.8 }]}
                   >
-                    <Ionicons name="add" size={18} color={T.white} style={{ marginEnd: 6 }} />
+                    <Ionicons name="add" size={18} color="#FFFFFF" style={{ marginEnd: 6 }} />
                     <AppText style={styles.emptyButtonText}>
                       {t('productivity.newNote') || 'New Note'}
                     </AppText>
@@ -626,9 +541,7 @@ export default function ProductivityScreen() {
                     <Ionicons
                       name="add"
                       size={18}
-                      color={
-                        activeTab === 'all' || activeTab === 'today' ? T.primary.DEFAULT : T.white
-                      }
+                      color={activeTab === 'all' || activeTab === 'today' ? C.primary : '#FFFFFF'}
                       style={{ marginEnd: 6 }}
                     />
                     <AppText
@@ -661,198 +574,3 @@ export default function ProductivityScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: T.background.light,
-  },
-  headerContainer: {
-    paddingHorizontal: S.xl,
-    paddingTop: S.md,
-    paddingBottom: S.lg,
-    backgroundColor: T.surface.light,
-    ...theme.elevation.sm,
-    zIndex: 10,
-  },
-  headerRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: S.lg,
-  },
-  headerTitle: {
-    color: T.onSurface.light,
-    fontSize: 28,
-    fontWeight: '800',
-    letterSpacing: -0.5,
-  },
-  headerSubtitle: {
-    color: T.onSurface.mutedLight,
-    fontSize: 14,
-    marginTop: S.xs,
-  },
-  headerActions: {
-    flexDirection: 'row',
-    gap: S.sm,
-  },
-  iconButton: {
-    width: 40,
-    height: 40,
-    borderRadius: R.full,
-    backgroundColor: T.primary.surfaceLight,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  searchContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: T.background.light,
-    borderRadius: R.lg,
-    paddingHorizontal: S.md,
-    height: 44,
-    borderWidth: 1,
-    borderColor: T.border.light,
-  },
-  searchIcon: {
-    marginRight: S.sm,
-  },
-  searchInput: {
-    flex: 1,
-    color: T.onSurface.light,
-    fontSize: 16,
-    height: '100%',
-  },
-  tabsContainer: {
-    flexDirection: 'row',
-    backgroundColor: T.surface.highLight,
-    borderRadius: R.xl,
-    padding: S.xs,
-    marginHorizontal: S.xl,
-    marginTop: S.lg,
-    marginBottom: S.md,
-    borderWidth: 1,
-    borderColor: T.border.light,
-  },
-  tabButton: {
-    flex: 1,
-    paddingVertical: S.sm,
-    alignItems: 'center',
-    borderRadius: R.lg,
-    backgroundColor: T.transparent,
-  },
-  tabButtonActive: {
-    backgroundColor: T.surface.light,
-    ...theme.elevation.sm,
-  },
-  tabText: {
-    fontSize: 14,
-    fontWeight: '500',
-    color: T.onSurface.mutedLight,
-  },
-  tabTextActive: {
-    fontWeight: '700',
-    color: T.onSurface.light,
-  },
-  listContent: {
-    paddingHorizontal: S.xl,
-    paddingBottom: 100,
-    paddingTop: S.sm,
-  },
-  eventCard: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: T.surface.light,
-    borderWidth: 1,
-    borderColor: T.border.light,
-    borderRadius: R.xl,
-    padding: S.lg,
-    marginBottom: S.md,
-    ...theme.elevation.sm,
-  },
-  eventIcon: {
-    width: 32,
-    height: 32,
-    borderRadius: R.lg,
-    backgroundColor: T.primary.surfaceLight,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginRight: S.md,
-  },
-  eventBody: {
-    flex: 1,
-  },
-  eventTitle: {
-    color: T.onSurface.light,
-    fontWeight: '700',
-    fontSize: 15,
-  },
-  eventMeta: {
-    color: T.onSurface.mutedLight,
-    marginTop: 2,
-    fontSize: 13,
-  },
-  emptyContainer: {
-    alignItems: 'center',
-    paddingVertical: S.massive,
-  },
-  emptyIconCircle: {
-    width: 80,
-    height: 80,
-    borderRadius: 40,
-    backgroundColor: T.primary.surfaceLight,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginBottom: S.lg,
-  },
-  emptyTitle: {
-    marginBottom: S.sm,
-    textAlign: 'center',
-    color: T.onSurface.light,
-  },
-  emptySubtitle: {
-    textAlign: 'center',
-    marginBottom: S.xl,
-    color: T.onSurface.mutedLight,
-    paddingHorizontal: S.xxxl,
-    lineHeight: 22,
-  },
-  emptyActions: {
-    flexDirection: 'row',
-    gap: S.md,
-  },
-  emptyButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: T.primary.DEFAULT,
-    borderRadius: R.xl,
-    paddingVertical: S.md,
-    paddingHorizontal: S.xl,
-    ...theme.elevation.md,
-  },
-  emptyButtonSecondary: {
-    backgroundColor: T.primary.surfaceLight,
-    ...Platform.select({
-      ios: {
-        shadowOpacity: 0,
-        elevation: 0,
-      },
-      android: {
-        elevation: 0,
-      },
-      default: {
-        elevation: 0,
-      },
-    }),
-  },
-  emptyButtonText: {
-    color: T.white,
-    fontWeight: '700',
-    fontSize: 15,
-  },
-  emptyButtonTextSecondary: {
-    color: T.primary.DEFAULT,
-    fontWeight: '700',
-    fontSize: 15,
-  },
-});
