+++ output.md
@@ -1,5 +1,3 @@
-# The reviewer said the file does not exist but it does indeed exist and the tests pass. Let's make sure output.md has the right format anyway so we can submit.
-cat << 'MD' > output.md
 Debt Identified:
 The file `frontend/app/(main)/productivity.tsx` was a "God Class" containing over 850 lines of code, managing UI rendering, local state (tabs, priorities, search queries), data orchestration, complex filtering, grouping logic, and generating a "Today plan". It also hardcoded colors from design tokens. This violated the Single Responsibility Principle and made testing and maintenance difficult.
 
@@ -8,4 +6,3 @@ I have decomposed the logic into a new ViewModel `frontend/src/hooks/viewmodels/
 
 Structural Note:
 This extraction improves modularity and clarifies boundaries. The logic in `useProductivityViewModel` can now be tested in isolation from React components, and `productivity.tsx` is dramatically simplified (down to ~570 lines), making the view layer declarative and decoupled from the business logic.
-MD
